feat(testing): Use the Jest testing framework and switch to GitHub Actions (#254)
* chore: use the jest testing framework

* chore: switch to github actions

The NPM_TOKEN secret should be set in an environment called "publish"
