+++ packages/adapters/.eslintrc.json
@@ -2,7 +2,8 @@
   "env": {
   "browser": true,
     "node": true,
-    "es6": true
+    "es6": true,
+    "jest": true
   },
   "extends": [
     "eslint:recommended"
