+++ packages/adapters/.github/workflows/tests.yml
@@ -1,0 +1,24 @@
+name: Run tests
+
+on:
+  - push
+  - pull_request
+
+jobs:
+  test:
+    runs-on: ubuntu-20.04
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v2
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v1
+        with:
+          node-version: 16
+
+      - name: Install packages
+        uses: bahmutov/npm-install@v1
+
+      - name: Run tests
+        run: npm run test
