+++ packages/adapters/package.json
@@ -8,7 +8,7 @@
     "example": "examples"
   },
   "scripts": {
-    "test": "node test/tests.js",
+    "test": "jest",
     "build": "rollup -c",
     "build:examples": "npm run build && npx cpx 'build/**/*.{js,map}' examples/js",
     "start": "rollup -c -w",
@@ -25,16 +25,16 @@
   },
   "homepage": "https://github.com/dcmjs-org/dcmjs#readme",
   "devDependencies": {
-    "@babel/core": "^7.8.4",
-    "@babel/preset-env": "^7.8.4",
+    "@babel/core": "^7.17.8",
+    "@babel/preset-env": "^7.16.11",
     "@babel/register": "^7.13.0",
     "acorn": "^7.1.0",
     "acorn-jsx": "^5.2.0",
     "babel-eslint": "^10.1.0",
-    "chai": "^4.2.0",
     "eslint": "^5.16.0",
     "follow-redirects": "^1.10.0",
     "husky": "^1.3.1",
+    "jest": "^27.5.1",
     "lint-staged": "^8.2.1",
     "prettier": "^1.19.1",
     "rollup": "^1.31.1",
@@ -48,8 +48,7 @@
     "xml2js": "^0.4.23"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.8.3",
-    "@babel/runtime": "^7.8.4",
+    "@babel/runtime-corejs2": "^7.17.8",
     "gl-matrix": "^3.1.0",
     "lodash.clonedeep": "^4.5.0",
     "loglevelnext": "^3.0.1",
