--- packages/adapters/src/utilities/TID1500/TID1500MeasurementReport.test.js
@@ -1,2 +1,0 @@
-// Load the file with DCMJS
-// Create a report
