+++ packages/adapters/test/adapters.test.js
@@ -1,0 +1,1 @@
+it("No tests yet", () => {});
