+++ packages/adapters/test/sr.test.js
@@ -1,0 +1,1 @@
+it("No tests yet", () => {})
