--- packages/adapters/test/test_DICOMWEB.js
@@ -1,5 +1,0 @@
-
-exports.test = () => {
-  console.log('no tests yet');
-}
-
