--- packages/adapters/test/test_adapters.js
@@ -1,5 +1,0 @@
-
-exports.test = () => {
-  console.log('no tests yet');
-}
-
