--- packages/adapters/test/test_sr.js
@@ -1,5 +1,0 @@
-
-exports.test = () => {
-  console.log('no tests yet');
-}
-
